+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.57.0 (unreleased)
 * Enable watch mode when code-splitting ([#2035](https://github.com/rollup/rollup/pull/2035))
 * Option to add correct `.toString` tags to namespaces ([#2041](https://github.com/rollup/rollup/pull/2041))
+* Option and scripts to display performance timings ([#2045](https://github.com/rollup/rollup/pull/2045))
 * Fixes for exported or early accessed namespaces + improved namespace indentation ([#2041](https://github.com/rollup/rollup/pull/2041))
 * Include missing TypeScript dependencies ([#1965](https://github.com/rollup/rollup/pull/1965))
 * [BREAKING] Revert class id preservation from #2025 ([#2048](https://github.com/rollup/rollup/pull/2048))